correct

+++ app/Vkontakte/Bot/BotCommandMethod.php
@@ -67,7 +67,10 @@ class BotCommandMethod
             $this->welcomeInviteMessageUser();
          }
       } else {
-         Chats::query()->where('question', $cache['question'])->delete();
+         if (Chats::query()->where('question', $cache['question'])->exists()) {
+            Chats::query()->where('question', $cache['question'])->delete();
+         }
+
          ChatQuestion::query()->create([
             'question' => $cache['question'],
             'answer' => $this->messageText
